Refactor token verification into an interface
